+++ R/html_method.R
@@ -1,111 +1,98 @@
 #' @include partition_class.R partitionBundle_class.R
 NULL
 
-#' restore fulltext as html
+#' Generate html from object.
 #' 
-#' @param object the object
-#' @param cqp logical
-#' @param cpos logical
-#' @param verbose logical
+#' Prepare a html document to inspect the full text.
+#' 
+#' @param object the object the fulltext output will be based on
+#' @param meta metadata for output, if NULL (default), the s-attributes defining
+#'   a partition will be used
+#' @param cpos logical, if TRUE (default), all tokens will be wrapped by 
+#'   elements with id attribute indicating corpus positions
+#' @param verbose logical, whether to be verbose
 #' @param filename the filename
-#' @param cutoff maximum number of tokens to decode from token stream
+#' @param cutoff maximum number of tokens to decode from token stream, passed
+#'   into \code{as.markdown}
 #' @param type the partition type
-#' @param i to be checked
-#' @param ... further parameters
+#' @param i if object is a \code{kwic}-object, the index of the concordance for
+#'   which the fulltext is to be generated
+#' @param ... further parameters that are passed into \code{as.markdown}
 #' @rdname html-method
 #' @aliases show,html-method
+#' @examples
+#' use("polmineR")
+#' P <- partition("REUTERS", places = "argentina")
+#' H <- html(P)
+#' H # show full text in viewer pane
+#' 
+#' # html-method can be used in a pipe
+#' if (require("magrittr")){
+#'   partition("REUTERS", places = "argentina") %>% html()
+#' }
+#' 
+#' # use html-method to get from concordance to full text
+#' K <- kwic("REUTERS", query = "barrels")
+#' html(K, i = 1)
+#' html(K, i = 2)
+#' for (i in 1:length(K)) {
+#'   show(html(K, i = i))
+#'   if (interactive()){
+#'     userinput <- readline("press 'q' to quit or any other key to continue")
+#'     if (userinput == "q") break
+#'   }
+#' }
 setGeneric("html", function(object, ...){standardGeneric("html")})
 
 
 #' @rdname html-method
 setMethod("html", "character", function(object){
-  if (requireNamespace("markdown", quietly = TRUE)){
-    mdFilename <- tempfile(fileext = ".md")
-    htmlFile <- tempfile(fileext = ".html")
-    cat(object, file = mdFilename)
-    markdown::markdownToHTML(file = mdFilename, output = htmlFile)  
-  } else {
+  if (!requireNamespace("markdown", quietly = TRUE)){
     stop("package 'markdown' is not installed, but necessary for this function")
   }
+  mdFilename <- tempfile(fileext = ".md")
+  htmlFile <- tempfile(fileext = ".html")
+  cat(object, file = mdFilename)
+  markdown::markdownToHTML(file = mdFilename, output = htmlFile)  
   htmlFile
 })
 
 
 
-#' @param meta metadata for output
-#' @param highlight list with regex to be highlighted
-#' @param tooltips tooltips for highlighted text
 #' @rdname html-method
 #' @exportMethod html
 #' @docType methods
 setMethod(
   "html", "partition",
-  function(
-    object, meta = NULL,
-    highlight = list(),
-    cqp = FALSE, tooltips = NULL, cpos = FALSE, verbose = FALSE, cutoff = NULL, ...
-    ){
-  if (requireNamespace("markdown", quietly = TRUE) && requireNamespace("htmltools", quietly = TRUE)){
+  function(object, meta = NULL, cpos = TRUE, verbose = FALSE, cutoff = NULL, ...){
+    if (!requireNamespace("markdown", quietly = TRUE) && requireNamespace("htmltools", quietly = TRUE)){
+      stop("package 'markdown' is not installed, but necessary for this function")
+    }
     if (is.null(meta)) meta <- names(object@sAttributes)
     if (all(meta %in% sAttributes(object)) != TRUE) warning("not all sAttributes provided as meta are available")
     
     .message("generating markdown", verbose = verbose)
-    if (any(cqp)) cpos <- TRUE
     markdown <- as.markdown(object, meta = meta, cpos = cpos, cutoff = cutoff, ...)
-    .message("markdown to html", verbose = verbose)
-    if (is.null(tooltips)){
-      htmlDoc <- markdown::markdownToHTML(text = markdown)
-    } else {
-      markdown_css <- scan(
-        getOption("markdown.HTML.stylesheet"),
-        what = "character", sep = "\n", quiet=TRUE
-        )
-      tooltips_css <- scan(
-        system.file("css", "tooltips.css", package="polmineR"),
-        what = "character", sep = "\n", quiet = TRUE
-        )
-      css <- paste(c(markdown_css, tooltips_css), collapse = "\n")
-      htmlDoc <- markdown::markdownToHTML(text = markdown, stylesheet = css)
-    }
     
-    if (length(highlight) > 0) {
-      if (length(cqp) == 1){
-        if (cqp == FALSE){
-          .message("highlighting regular expressions", verbose = verbose)
-          htmlDoc <- highlight(htmlDoc, highlight = highlight, tooltips = tooltips)
-        } else if (cqp){
-          .message("highlighting CQP queries", verbose = verbose)
-          htmlDoc <- highlight(object, htmlDoc, highlight = highlight, tooltips = tooltips)
-        }
-      } else if (length(cqp) == length(highlight)){
-        if (any(!cqp)){
-          htmlDoc <- highlight(
-            htmlDoc,
-            highlight = highlight[which(cqp == FALSE)],
-            tooltips = tooltips[which(cqp == FALSE)]
-          )
-        }
-        if (any(cqp)){
-          htmlDoc <- highlight(
-            object, htmlDoc,
-            highlight = highlight[which(cqp == TRUE)],
-            tooltips = tooltips[which(cqp == TRUE)]
-          )
-        }
-      } else {
-        message("length of cqp needs to be 1 or identical with the length of highlight")
-      }
-    }
+    .message("markdown to html", verbose = verbose)
+    # the css for tooltips is included in the html document by default,
+    # even if this causes some (minimal) overhead.
+    # The default stylesheet (markdown.css) needs to be included explicitly,
+    # so that it is not lost.
+    css <- paste(c(
+      readLines(getOption("markdown.HTML.stylesheet")),
+      readLines(system.file("css", "tooltips.css", package = "polmineR"))
+      ), collapse = "\n", sep = "\n"
+    )
+    htmlDoc <- markdown::markdownToHTML(text = markdown, stylesheet = css)
     htmlDocFinal <- htmltools::HTML(htmlDoc)
-  } else {
-    stop("package 'markdown' is not installed, but necessary for this function")
+    htmlDocFinal
   }
-  htmlDocFinal
-})
+)
 
 #' @docType methods
 #' @rdname html-method
-setMethod("html", "partitionBundle", function(object, filename = c(), type="debate"){
+setMethod("html", "partitionBundle", function(object, filename = c(), type = "debate"){
   markdown <- paste(lapply(object@objects, function(p) as.markdown(p, type)), collapse="\n* * *\n")
   markdown <- paste(
     paste('## Excerpt from corpus', object@corpus, '\n* * *\n'),